Support uint64 in the deterministic Miller-Rabin algorithm

+++ src/primes.h
@@ -7,7 +7,7 @@ bool is_prime_naive(uint64_t x);
 // Deterministic Miller-Rabin algorithm. Selects the witnesses as per
 // https://en.wikipedia.org/wiki/Miller%E2%80%93Rabin_primality_test.
 // Becomes faster than the naive algorithm for x > 200`000`000.
-bool is_prime_miller_rabin(uint32_t x);
+bool is_prime_miller_rabin(uint64_t x);
 
 // Runs the Miller-Rabin algorithm `attempts` times.
 // Tells whether x is a prime with ≤(1/4)^attempts false positive and zero
